Refactoring for readability.

+++ include/xtal/process/any.c++
@@ -18,7 +18,7 @@ TEST_CASE("xtal/process/any.hpp: lifting")
 	auto const f = let_f([](XTAL_DEF... xs) XTAL_0FN_(XTAL_REF_(xs) +...+ 0));
 	REQUIRE(10 == f.template method<>(1, 2, 3, 4));
 	REQUIRE(10 == f(1, 2, 3, 4));
-	REQUIRE(10 == f.reify()(1, 2, 3, 4));
+	REQUIRE(10 == f.reify() (1, 2, 3, 4));
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -71,11 +71,11 @@ TEST_CASE("xtal/process/any.hpp: attribute efflux operator")
 void attribute_influx_method__test(auto z)
 {
 	auto &o = z.template self<bias_t>();
-	REQUIRE(-1 == z.influx(control::start_t()));                           // unrecognized
-	REQUIRE( 0 == z.influx(bias_t(0.0))); REQUIRE(0.0 == (float) o.head());// unchanged
-	REQUIRE( 1 == z.influx(bias_t(1.0))); REQUIRE(1.0 == (float) o.head());// changed
-	REQUIRE( 1 == z.influx(bias_t(2.0))); REQUIRE(2.0 == (float) o.head());// changed
-	REQUIRE( 1 == z.influx(bias_t(3.0))); REQUIRE(3.0 == (float) o.head());// changed
+	REQUIRE(-1 == (int) z.influx(control::start_t()));                           // unrecognized
+	REQUIRE( 0 == (int) z.influx(bias_t(0.0))); REQUIRE(0.0 == (float) o.head());// unchanged
+	REQUIRE( 1 == (int) z.influx(bias_t(1.0))); REQUIRE(1.0 == (float) o.head());// changed
+	REQUIRE( 1 == (int) z.influx(bias_t(2.0))); REQUIRE(2.0 == (float) o.head());// changed
+	REQUIRE( 1 == (int) z.influx(bias_t(3.0))); REQUIRE(3.0 == (float) o.head());// changed
 	REQUIRE(13.0 == (float) z(1.0, 2.0, 3.0, 4.0));
 }
 TEST_CASE("xtal/process/any.hpp: attribute influx method")
@@ -88,11 +88,11 @@ TEST_CASE("xtal/process/any.hpp: attribute influx method")
 void attribute_efflux_method__test(auto z)
 {
 	auto &o = z.template self<bias_t>();
-	REQUIRE(-1 == z.efflux(control::start_t()));                           // unrecognized
-	REQUIRE( 0 == z.efflux(bias_t(0.0))); REQUIRE(0.0 == (float) o.head());// unchanged
-	REQUIRE( 1 == z.efflux(bias_t(1.0))); REQUIRE(1.0 == (float) o.head());// changed
-	REQUIRE( 1 == z.efflux(bias_t(2.0))); REQUIRE(2.0 == (float) o.head());// changed
-	REQUIRE( 1 == z.efflux(bias_t(3.0))); REQUIRE(3.0 == (float) o.head());// changed
+	REQUIRE(-1 == (int) z.efflux(control::start_t()));                           // unrecognized
+	REQUIRE( 0 == (int) z.efflux(bias_t(0.0))); REQUIRE(0.0 == (float) o.head());// unchanged
+	REQUIRE( 1 == (int) z.efflux(bias_t(1.0))); REQUIRE(1.0 == (float) o.head());// changed
+	REQUIRE( 1 == (int) z.efflux(bias_t(2.0))); REQUIRE(2.0 == (float) o.head());// changed
+	REQUIRE( 1 == (int) z.efflux(bias_t(3.0))); REQUIRE(3.0 == (float) o.head());// changed
 	REQUIRE(13.0 == (float) z(1.0, 2.0, 3.0, 4.0));
 }
 TEST_CASE("xtal/process/any.hpp: attribute efflux method")
